requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
requests
masai_framework==0.5.2
